modularize typeDefs
