+++ src/graphql/schema/global/index.js
@@ -1,0 +1,8 @@
+/*eslint-disable */
+const typeDefs = require('./typeDefs')
+// const resolvers = require('./resolvers')
+
+module.exports = {
+  typeDefs
+  // resolvers
+}
