+++ src/graphql/schema/global/typeDefs/index.js
@@ -1,0 +1,6 @@
+/*eslint-disable */
+const { Query } = require('./entryPoints')
+const { scalars } = require('./scalars')
+const { spaceX } = require('./spaceX')
+
+module.exports = [Query, scalars, spaceX]
