+++ src/graphql/schema/global/typeDefs/scalars.js
@@ -1,0 +1,11 @@
+/*eslint-disable */
+const { gql } = require('apollo-server')
+
+const scalars = gql`
+  scalar Date
+  scalar ObjectID
+`
+
+module.exports = {
+  scalars
+}
