+++ src/graphql/schema/typeDefs.js
@@ -1,30 +1,8 @@
 /*eslint-disable */
-const { gql } = require('apollo-server')
+const { typeDefs: global } = require('./global')
 const { typeDefs: capsule } = require('./capsule')
 const { typeDefs: core } = require('./core')
 
-const Scalars = gql`
-  scalar Date
-  scalar ObjectID
-`
-const Query = gql`
-  type Query {
-    _empty: String
-  }
-`
-const Mutation = gql`
-  type Mutation {
-    _empty: String
-  }
-`
-const Subscription = gql`
-  type Subscription {
-    _empty: String
-  }
-`
-
-const globals = [Scalars, Query, Mutation, Subscription]
-
-const typeDefs = [...globals, capsule, core]
+const typeDefs = [...global, capsule, core]
 
 module.exports = typeDefs
